Close and unmount filesystem if an error occurs
Previously an error would occur on boot since the file is being read
which makes sense since the file would not always exist. However since
the mutex was never unlocked it would be impossible for the download
module to download and create the file.

We should probably look into a more elegant way of handling this.